+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/util/modifiable/MPRModifiableValue.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @JsonAdapter(MPRModifiableValue.Serializer.class)
 public class MPRModifiableValue extends MPRModifiable {
-	protected final Double value;
+	public final Double value;
 
 	public static final MPRModifiableValue ZERO = new MPRModifiableValue(0d);
 	public static final MPRModifiableValue ONE = new MPRModifiableValue(1d);